+++ examples/localasset.jl
@@ -1,0 +1,8 @@
+function helloWorld2()
+  println("HelloWorld1 from localasset")
+end
+
+function squareDiv(a::Float, b::Float)
+  println("squareDiv from localasset")
+  return (a*a)/(b*b)
+end