+++ nimjl/glucose.nim
@@ -9,11 +9,30 @@ import ./private/jlcores
 
 type Julia* = object
 
+template Embed*(jl: type Julia, body: untyped) =
+  JlEmbed(body)
+
+proc embedFile*(jl: type Julia, fname: static[string]) =
+  jlEmbedFile(fname)
+
+proc embedDir*(jl: type Julia, dirname: static[string]) =
+  jlEmbedDir(dirname)
+
 proc init*(jl: type Julia) =
+  jlVmInit()
+
+template init*(jl: type Julia, body: untyped) =
+  JlEmbed(body)
   jlVmInit()
 
 proc exit*(jl: type Julia, exitcode: int = 0) =
   jlVmExit(exitcode.cint)
+
+proc useModule*(jl: type Julia, modname: string) =
+  jlUseModule(modname)
+
+proc includeFile*(jl: type Julia, fname: string) =
+  jlInclude(fname)
 
 # macro loadModule*(jl: type Julia, modname: untyped) =
 # TODO generate a proc ``modname`` that returns module